+++ CHANGES.rst
@@ -4,6 +4,7 @@ Version 1.0 (dev)
 * Added multilingual support, using django-parler_.
 * Added frontend media support.
 * Added "Open in new window" option for the "picture" plugin.
+* Added ``HttpRedirectRequest`` exception and ``HttpRedirectRequestMiddleware``.
 * Content items are prefixed with "content:" during syncdb, a ``prefix_content_item_types`` management command can be run manually too.
 * **API Change:** Renamed template tag library ``placeholder_tags`` to ``fluent_contents_tags`` (the old name still works).
 * **API Change:** ``render_placeholder()`` and ``render_content_items()`` return a ``ContentItemOutput`` object, which can be treated like a string.